move x11.c into old folder
